Fix various UI bugs
Fixed the divider between time/date and media player showing even when the media player was not visible. Fixed an issue where favourited apps may not be right-aligned on some devices. Limited the amount of columns on fixed mode to 3. Design and space improvements for smaller devices and landscape mode.